Improved Navigation - added support for passing arguments to form from general Add button in list (#47)
* moved some logic inside panel component instead of single pages to support same component nested

* moved some route logic to manage preload of routes

* fixed breadcrumbs navigation

* fixed form params

* added support for passing params from where filter into form

+++ src/app/panel/services/setup.service.ts
@@ -1,28 +1,30 @@
 import {Injectable} from '@angular/core';
 import {ApiService} from '../../api/api.service';
-import {Router, Routes} from '@angular/router';
+import {Router} from '@angular/router';
 import {environment} from '../../../environments/environment';
 import {DashboardPageComponent} from '../pages/dashboard-page/dashboard-page.component';
 import {TablePageComponent} from '../pages/table-page/table-page.component';
 import {FormPageComponent} from '../pages/form-page/form-page.component';
 import {ProfilePageComponent} from '../pages/profile-page/profile-page.component';
-import {NotfoundPageComponent} from '../pages/notfound-page/notfound-page.component';
-import {UtilsService} from '../../services/utils.service';
 import {UserService} from '../../auth/services/user.service';
 import {LanguageService} from './language.service';
+import {MenuService} from './menu.service';
+import {NotfoundPageComponent} from '../pages/notfound-page/notfound-page.component';
 
 const TYPES = {
     profile: ProfilePageComponent,
     dashboard: DashboardPageComponent,
     table: TablePageComponent,
     form: FormPageComponent,
-    notFound: NotfoundPageComponent
 };
 
 @Injectable()
 export class SetupService {
 
+    private _lastRouteLoading: Date;
+
     constructor(private _router: Router,
+                private _menuService: MenuService,
                 private _userService: UserService,
                 private _apiService: ApiService,
                 private _languageService: LanguageService) {
@@ -30,22 +32,32 @@ export class SetupService {
 
     public setup(): Promise<any> {
         return new Promise<any>((resolve, reject) => {
+
+            if(this._lastRouteLoading == null || Date.now() - this._lastRouteLoading.getMilliseconds() < 10000) {
             this._apiService.get(environment.api.setupEndpoint)
                 .then((data) => {
+
+                   this._lastRouteLoading = new Date();
 
                     if ('contentLanguages' in data) {
                         this._languageService.setContentLanguages(data['contentLanguages']);
                     }
 
                     this.loadRoutes(data);
-                    const menu = this.prepareMenu(data);
-                    resolve(menu);
+                    console.log("CARICAMENTO ROTTE...");
+                    this._menuService.prepareMenu(data);
+                    resolve(true);
                 })
                 .catch((error) => {
                     this._userService.removeToken();
                     this._userService.removeUser();
-                    reject();
+                    resolve(false);
                 });
+            } else {
+                console.log("SALTO CARICAMENTO ROTTE...");
+                resolve(true);
+            }
+
         });
     }
 
@@ -71,7 +83,8 @@ export class SetupService {
 
         const routerConfig = this._router.config;
 
-        const routes = [];
+        const routes = [ {path: '404', component: NotfoundPageComponent}];
+
 
         for (const item of this.remapRoutesData(data)) {
 
@@ -85,51 +98,8 @@ export class SetupService {
             }
         }
 
-        /** Add not found page */
-        routes.push({
-            path: '**',
-            component: TYPES.notFound
-        });
-
         routerConfig[0].children = routes;
         this._router.resetConfig(routerConfig);
     }
 
-    private remapMenu(data: any): any[] {
-        let menu = [];
-
-        for (const item of ('pages' in data) ? data.pages : data) {
-            if ('type' in item) {
-                // Is a component
-                if (item.type === 'group') {
-                    if (item.params.menu.sidebar) {
-                        if ('children' in item) {
-                            if (item.params.menu) {
-                                item.params.menu['children'] = this.remapMenu(item.children);
-                                menu.push(item.params.menu);
-                            }
-                        }
-                    }
-                } else {
-                    if (item.params.menu) {
-                        if (item.params.menu.sidebar === true) {
-                            item.params.menu.path = item.path;
-                            menu.push(item.params.menu);
-                        }
-                    }
-                }
-            } else if ('children' in item) {
-                menu = menu.concat(this.remapMenu(item.children));
-            }
-        }
-
-        return menu;
-    }
-
-    private prepareMenu(data: any): any[] {
-        let menu = this.remapMenu(data);
-        menu = UtilsService.sortByKey(menu, 'order');
-
-        return menu;
-    }
 }
